css: put float64 directly in parse tree
This is a questionable decision. The CSS specification does not seem
to place any size limit on numeric values. But that also means the
previous uint64 was wrong: float64 may not capture even the float
values we see.

So either we give up on numeric parsing altogether in the parser,
which is tempting and I might do in the future, or we accept that
our parser is only going to handle a subset of possible CSS.