css: format functions and unicode ranges
Add some TODOs for function representation.
It could be improved beyond the partial-parsing we are doing now,
but it should be thought about in the context of a full sheet
parser, rather than just the decl parsing we are doing now.
(Should Value be reused, and should it contain a slice of Values?)

Combine the parse and format tests where they can be.