spillbox: move blobs table to its own database
This keeps the main user database small.
So small that there is little need to do incremental backups
of the main database.
For live server backups, the SQLite online backup API can be
used to grab a copy of each main user database, and the blobs
can be incrementally backed up matched by hash.